Delete update task and move logger to run earlier
Somehow in MacOS the logs less stuff by default.
