+++ src/main/java/de/danielluedecke/zettelkasten/CErrorLog.java
@@ -113,7 +113,7 @@ public class CErrorLog extends javax.swing.JDialog {
                             + "------------------------------"
                             + System.lineSeparator()
                             + System.lineSeparator()
-                            + mainframe.in_memory_session_log.toString());
+                            + ZettelkastenApp.getApplication().getCurrentSessionLogs().toString());
                     jTextArea2.setCaretPosition(0);
                 }
             }
